Add files via upload
The goal of this project is to create a small compiler for a subset of the language C, called "C--", to an intermediate language also a subset of C called "intC" with some optimizations. The translation from C-- to intC requires introducing new variables to comply with three-address instructions. The compiler can introduce an unlimited number of variables. The generated new variables must start with an underscore character and must be declared as local variables with the appropriate type. On the other hand, since multidimensional arrays are not allowed in intC, C-- multidimensional arrays must be linearized.